refactor: make INumber conform to v-model
modelValue and update:modelValue are actually what it's supposed to be.
TIL: https://vuejs.org/guide/components/events.html#usage-with-v-model

So IString works to some extent, but it's not getting value from outside,
but it's fully controlled by the internal ref (because modelValue is only used
as defautl value). We should modify IString to work this way later too.