fix: default to null for wallet in edit space modal
Currently when opening edit space modal for space without
treasury configured it will show wallet address field, even though
it shouldn't do that.

That's because we are defaulting to '' and when decoding we might put
undefined values into them as well.

This PR changes those fields to default to null.