fix: remove subtree option for BlockInfiniteScroller
BlockInfiniteScroller used to wrap entire table but it wouldn't trigger
endReached event because table itself didn't have any direct
children added to it, so subtree property was added, but it caused
different issue as now table's last element is always visible on screen
(tbody) so it triggers endReached event all the time.

subtree property is not good solution and we should always wrap just the
elements we want to load.